refactor: change background image to be the same for all pages
