+++ src/pages/dashboard/Agenda.jsx
@@ -3,7 +3,7 @@ import AgendaList from "../../components/dashboard/AgendaList";
 
 function Agenda() {
   return (
-    <div className={styles.agendaPage}>
+    <div className={styles.backgroundPages}>
       <div className={`container mt-5 ${styles.agendaPanel}`}>
         <AgendaList />
       </div>
