+++ src/pages/dashboard/Home.jsx
@@ -3,7 +3,7 @@ import styles from "./Dashboard.module.css";
 function Home() {
   return (
     <div>
-      <div className={`${styles.homePage}`}>
+      <div className={`${styles.backgroundPages}`}>
         
       </div>
     </div>
