+++ src/pages/farms/Farms.jsx
@@ -1,9 +1,9 @@
 import FarmList from "../../components/farms/FarmList";
-import styles from "./Farms.module.css";
+import styles from "../dashboard/Dashboard.module.css";
 
 function Farms() {
   return (
-    <div className={`pb-5 ${styles.farmPage}`}>
+    <div className={`pb-5 ${styles.backgroundPages}`}>
       <div className="container mt-5">
         <FarmList />
       </div>
