+++ src/pages/farms/Farms.module.css
@@ -1,19 +1,3 @@
-.farmPage{
-    background-image: url('/img/imgVerFincas.png');
-    background-size: cover;
-    background-position: center;
-    min-height: 100vh;
-    
-}
-
-.farmForm{
-    background-image: url('/img/imgAddFinca.png');
-    background-size: cover;
-    background-position: center;
-    min-height: 100vh;
-
-}
-
 .formPanel{
     background-color: rgba(236, 236, 236, 0.8);
     border-radius: 10px;
