+++ src/pages/suppliers/AddSupplier.jsx
@@ -1,9 +1,9 @@
 import RegisterForm from "../../components/forms/RegisterForm";
-import styles from "./Suppliers.module.css";
+import styles from "../dashboard/Dashboard.module.css";
 
 function AddSupplier() {
   return (
-    <div className={`pb-5 ${styles.addSupplierPage}`}>
+    <div className={`pb-5 ${styles.backgroundPages}`}>
       <RegisterForm type={"supplier"} />
     </div>
   );
